factor out add genes to own component
